Refactor and move the class package Prior to this change, class was a separate package creating, deleting classes and reading, writing to the db by itself.
This change adds the class package to the storage package. Now the class
struct is not anymore 'class.Class' but 'storage.Class'. Now the class
only represents an actual class. The actual reading and writing to the
database is now handeld by the Storage interface and the cli interaction
is handeld in the cmd package.