Refactor and move the project package
Prior to this change, project was a separate package creating
project and reading, writing them to the db by itself.

This change adds the project package to the storage package. Now the
project struct is not anymore 'project.Project' but 'storage.Project'.
The project file only represents an actual project. The actual writing
to the database is handeld by the storage interface and the reading is
done through a class. Atm the project class doesn't execute any reading
on the database. It reads the data from the given class which the
project is meant to be based on. The cli interaction is handeld in the
cmd package.