Refactor the storage interface and sqlite package Prior to this change, the sqlite package didn't contain much functionality. Pretty much an empty and dead file.
This change fills the sqlite package with all the code that's necessary
to read and write classes and projects from and to the sqlite3 database.
All the database interactions that were previously done other packages
are now handeld in this package. The Storage interface serves as
representation for the necessary behaviour of a storage related package.
It defines the necessary functions to save, load, remove classes and
projects.