Refactor cli interaction Prior to this change, the separate files were fully dependend on the data packages like class and project. Most of the cli output were created in these packages not by the cmd package.
This change comletely changes the way the cmd package works. Now cmd
file that represents a cli command merges the functionalities of the
class and project structures with the storage interface. CLI content is
only created here. Class and project represent the actual objects,
storage reads and writes to them and the cmd package now
represents the objects.