Add multitude of functions to manage projects and statuses
Prior to this change, the only project related function was the Create
function which obviously created a new project and tracked it in the
database.

This change adds a lot of functions to add, remove, update and list
projects and statuses. Will be used as base for coming project and
status related CLI commands.