Add function to load the ID of a project Since projects are best identified through their ID there should be a function that gets the ID for a given project.
This change adds a function that returns the ID of a project if the
project exists. The project is identified through its install path since
at one location only one project can exist but multiple projects might
share the same name. That's why the path and not the name is used to identify the project.