Add WantTo function
In the CLI its often needed to ask the user if he wants to actually
execute an action. 'Do you want to delete this and that? [y/N]'.

This change adds a function that takes the question as parameter and
runs a loop aslong as the user didn't press y or n. Depending on his
choice the function returns a boolean.