Add unique index for project_status table
Prior to this change, it was expected that the project_status table
would report a unique constraint violation when the user would try to
add a status title which already existed. Not the case.

This change adds the unique contraint on the title column in the
project_status table. Now status titles are unique.