Add function UpdateStatus
Prior to this change, I had to manually remove and readd a status if I
wanted to replace a status. Besides of the need of calling two functions
to replace a status it also messed up the projects that were assigned to
the status that was being deleted. When the status with ID 1 gets
deleted and the new status gets inserted, then the new status gets an ID
greater than 1. And projects that were created with the now deleted
status 1 now reference a non-existend status.

This change adds a function to simply update the already existing
status. When 'proji status add' detects that the status already exists
why should the existing status be removed? They have the same title. The
only difference might be the comment.