Add LoadStatusID function
Since the status is now actively used in the cli and can be added,
removed and modified, the IDs of statuses are frequently in demand.

This change adds a function to easily return the ID for a given status
title. It returns an error if the status does not exist.