Add the sub command 'add' to 'status'
Prior to this change, the user was not able to add personalized statuses
which would be a helpful feature.

This change adds the command 'proji status add' followed by statuses.
The app checks if a given status already exists, offers to update its
comment if so or adds the new status if not.