Add sub command 'ls' to 'status'
Prior to this change, the user was not able to see a list of available
statuses which is pretty much needed to remove statuses since the remove
command wants status IDs.

This change adds the sub command 'proji status ls' which will list all
statuses with their ID, title and comment nicely formatted in a table.