Replace regular vars with special sql vars
Prior to this change, the app might have failed when the scan of a query
result would have returned a null value. The regular go types can't
handle that.

This change replaces the the regular go types with their equivalent
null-types from the sql package. Now scan can set null values to the
variables.