Fix bug in script execution order
Prior to this change, scripts were selected from the storage ordered by
their sudo value and name. This led to an other order of execution of
the scripts than the user might have intended to. If I want to run the
git script before the virtualenv script then I'd add them to a class in
this order. But the prior query selected them in another order from the
storage, possibly leading to the virtualenv script running before git.

This change orders the select by ID. The ID represents the order in
which the user originally added the scripts.