Fix bug in output of exported file name
Prior to this change, the command 'class export' was intended to print
the name of the created export file on success but it printed an empty
string.

After debugging the bug it appeared that the variable intended to hold
the file name was accidentally overwritten with an empty string. Fix the
overwriting and now the file name is printed on success.