Add the initialization of the config path and storage service
Prior to this change, every command created its own connection to a
storage service. Alot of code repetition.

This change initializes the config path and the storage service once.
After the initialization the path and service are readable by all
members of the package cmd through the package wide unexported struct
env. In future it will be possible to simply change the storage type
since it's now defined once and not in every subcommand.