Implement new env struct
Prior to this change, each command created a storage service and called
helper functions to get the config path which ended in very repetetive
and error prone code.

This change removes those service setups and helper calls. Now every
command uses the service and config path from the env struct which gets
defined in the root file of the cmd package.