Add package item
Prior to this change, the items were stored in the storage package.
Items are class, status and project.

This change separates the items from the storage package. Class, project
and status are now no longer accessible as 'storage.*' but as 'item.*'.
Replace all occurences of old storage calls with calls to the item
package.