Remove error return value from New* functions
Prior to this change, the constructor functions returned an error which
was needed in the past because the functions had elements that could
fail. But those elements were removed in one of the latest commits and
because of that all constructor functions always return nil which makes
the need for the error value unnecessary.

This change removes the error return value from the definitions and from
the function calls.