Change path and name of binary files
Prior to this change, the binary files were created under bin/ and were
named after the os and architecture they were created for.

This change changes this behaviour. Now the files are stored in a folder
structure named after the os and architecture.

Old: proji-linux-amd64
New: linux/amd64/proji

This helps with the packaging process for the release versions.