Fix script execution order
Fix a bug where the execution order of class scripts was random. Class
scripts are meant to run in the order they were added. Due to the
internal functionality of go maps the order was sometimes random.

This change adds three new structs representing a folder, a file, and a
script. Previously, the attributes, files, and scripts of the structure
class were of the type map [string]string/bool. They are now replaced
by arrays of the representative structures. On the one hand, this fixes the
access order to the scripts, which is now ordered, and moreover, these
structures are more expressive than the map types.

Closes #59.