Add more example files
Prior to this change, there was only an example class config with a
bunch of annotations explaining how the config file is structured and
how to edit it.

This change adds two examples with actual configs for python and cpp.
And in addition two scripts that are usable by proji.