Add default class 'unknown'
Prior to this change, proji would struggle and report errors when a
class that was used before was deleted. When listing projects and one of
the projects was created with a class that does not exist anymore, proji
would break and report that a class with a certain id was not found.
That doesn't help the user at all.

This change adds the default class 'unknown' which will always have the
id 1 and is unremoveable. When there are projects with classes assigned
that don't exist anymore, proji will add the class 'unknown' to it.
Which fixes the break and gives the user an information on what's wrong.