Fix a bug when creating multiple projects
Prior to this change, the copying of templates would fail when you'd
want to create multiple projects at once. Proji would report that
certain templates could not be found. The path that proji reported was
the default template folder path but appended to itself multiple times.

This change fixes this bug by stopping the function copyTemplates from
appending the configpath to itself everytime a new project was created.
Now the function creates the path in a temporary variable and not
anymore in the struct member.