Update readme
Since the last update of the readme a lot of changes happened to the
program itself. The readme needed to be updated and extended.

This change updates the readme. It now describes the installation and
usage extensivly. Showing gifs of proji's functionality and providing
lots of informations about how proji works internally.