Remove the installation of shell completion
Prior to this change, the installer would try to install the shell
completion file. This process was not stable due to problems with
reading the fpath in a bash script. Switching to a zsh script is also
does not seem like an option because zsh might not be installed on some
systems.

This change removes the install process of the shell completion file
completely. There will be instructions to manually enable tab completion in the readme.