Fix a bug in sqlite_test
Recently there was a default class added to the database. The class
'unknown'. This class exists by default. The test added a couple of
classes to the database, loaded them from the database again and then
compared the lists lengths. They should be equal. The test didn't put
the default class into the calculation which made the list of loaded
classes always one element bigger than the original class list.

This change subtracts the number of default classes from the loaded list
before the comparison is done.