Fix a bug in sqlite test
The test was comparing two lists of classes element by element. The list
with the classes that were loaded from the database was always bigger by
one element due to the default class that was loaded. This caused a loop
that iterated over both arrays to crash with an out of bounce error.

This change skips the default class in the comparison loop which stops
it from stepping out of bounce.