DB - class table & project_status: Add new column is_default
Prior to this change, there was no specification if a class or status was default
or not. Proji comes with a couple of default classes and statuses which
can not be removed by the user. To identify internally if a class or
script was default you had to know either the name or id of it. Then you could
skip these classes in functions like RemoveAllClasses for example. This
is obviously not a good way of identifying default members. What if we
add new default classes and statuses. Everywhere in the could we would
need to add the new names and ids to the checklists to prevent these
members from being deleted for example.

This change adds the new column 'is_default' to the tables class and
project_status. The column will hold integer values which will always be
either 0 or 1. 1 meaning the class/status is default and 0 meaning it is
not. Additionally we added the 'IsDefault' member to the class and status struct.
IsDefault is of type boolean and will determine if a class/status is
default. Internally we can now easily check for default classes and
statuses. Strongly improved the maintainability and readability.

Closes #60.