Add conditional clean up and install to install script
Prior to this change, the installer didn't check if a proji binary
existed in the cwd. It just ran the install command no matter if the
binary existed or not. This didn't break anything but caused irritating
error messages when the binary didn't exist. Additionally the installer
didn't remove the duplicate binary after it was installed.

This change checks if there is a proji binary in the cwd and only when
there it the installer installs the binary and cleans it up afterwards.