cli: add two flags to 'class import' command
Add two flags to the 'class import command' as preperation for the
coming support of includeable directories. The two flags are 'config'
and 'directory'. With the 'config' flag the user will import class
configs as before and with the 'directory' flag the user will be able to
import a class from an existing directory. Proji will copy the structure
and content of the directory and create a class config based on it.