func: add a function that will auto pick a label based on the class name
The function will first try to split the class name by any of the following separators: '-', '_', '.',
' '. On success the label will exist out of the first letters of the
first 4 or less words that the name was splitted into.
If that fails the function will do the same for upper case chars in the
string.
If that fails aswell the function will create the label from the first,
mid and last element of the name.

Examples:
'myTestProject' -> 'mtp'
'proji-is-awesome' -> 'pia'
'myclass' -> 'mls'