fix: label and/or name could be empty on import
When importing a class from a config file the name and label fields
could be empty and proji would still import the class which obviously
could have led to problems.

Now the name and label fields are checked before the imported class is
saved. An error is returned if one of the strings is empty.

Fix #72