test: fix an export related test
The affected test was still expecting the old config name format, which
was only the file name. Since the last commit, proji allows to specify a
'target' flag in the 'class export' command. Therefore, the resulting
file name in the test is preceded by a path, while the test expects a
pure file name.

This fix prefixes the expected filename with the destination path.