test: remove gh action for testing go code
The action did not work as intended. It didn't seem to use the pushed
branch for testing purposes. When for example a function was assigned a
new argument in the latest commit on that branch, the action test failed,
claiming that the function has too many arguments. When investigating
the pushed branch in gh you could clearly see, that the function was
correctly updated.
For now disabling the gh action until a solution is found.