feat: add optional args to class scripts
Args will be an array of strings holding arguments that would usually be
passed to the scripts via cli. The args are optional; the array may be
empty.
The args support the keyword __PROJECT_NAME__. Specifying this keyword
in the argument list of a class script will cause proji to pass the name
of the current project to the script when creating a project based on
that class.