feat: add type attribute to class scripts
The type attribute defines the type of the script. The currently
available options are 'pre' and 'post'. Pre meaning that the script will
be run before proji creates subdirs and files and post meaning the
default behaviour of scripts; running after everything got set up by
proji.

Types are superior to execution numbers. A script of type 'post' with a
execution number that is lower than that of a script of type 'pre' will be
executed after that script.

Examples:

[[script]]
  name = "myScript1.sh"
  type = "post"
  execNumber = 1
  ...

[[script]]
  name = "myScript2.sh"
  type = "pre"
  execNumber = 1
  ...

-> myScript2.sh will run first

[[script]]
  name = "myScript1.sh"
  type = "post"
  execNumber = 1
  ...

[[script]]
  name = "myScript2.sh"
  type = "pre"
  execNumber = 8
  ...

-> myScript2.sh will run first

[[script]]
  name = "myScript1.sh"
  type = "pre"
  execNumber = 1
  ...

[[script]]
  name = "myScript2.sh"
  type = "pre"
  execNumber = 1
  ...

-> Error: Two scripts of the same type can't share the same execNumber

[[script]]
  name = "myScript1.sh"
  type = "pre"
  execNumber = 1
  ...

[[script]]
  name = "myScript2.sh"
  type = "pre"
  execNumber = 2
  ...

[[script]]
  name = "myScript3.sh"
  type = "post"
  execNumber = 1
  ...

[[script]]
  name = "myScript4.sh"
  type = "post"
  execNumber = 2
  ...

-> Execution order: myScript1 -> myScript2 -> myScript3 -> myScript4

Closes #70