feat: add 'remote-repo' flag to 'class import' command
The flag 'remote-repo' enables the user to create a class based on a
remote repository from platforms like github for example. The repo is
not cloned. Proji uses a crawler to identify all the files and folders
the project consists of.
The rest is pretty similar to the 'directory' flag. Proji will export a
class config based on the remote repo which the user can edit if needed
and then be imported.

Closes #81