refactor: optimize code and add live multiline terminal output
Commands that may be loading multiple items from storage were optimized.
Code is now less and easier to read. No noticeable performance
differences.
These commands are now able to live update the terminal output when
importing/exporting/removing... items from storage. This makes it easier
to follow the execution state of the command. Is it still loading? What
is it removing right now?