refactor: use storage service in env struct directly
The projiEnv struct holds a pointer to the storage service which is
responsible for managing the data flows. Previously a pointer to the
storage service was passed to every function in need which was very
repetetive.
Now the functions access the projiEnv struct directly, the passing of a
pointer to the functions is thus not necessary anymore.