refactor: replace web crawler with new repo package
Until this point a web crawler was responsible for fetching all the
files and folders of a remote repository. It was inconsistent, slow and
a pain to maintain. With the new repo package which uses the public APIs
of the code hosting platforms the time to import
https://github.com/golang/go was reduced from 25 seconds to 2-3 seconds.
Additionally the results are 100% consistent and it is super easy to
maintain and extend.

Closes #96