refactor: make regex more precise
add webplatform prefix to regex. Previously the regex would only filter
the path after the domain name. When a github url was passed to gitlabs
New() function for example, the function would have parsed the url
anyway. Now the regex is prefixed with the domain of the respective
platform resulting in cleaner results.