test: fix a mac specific bug in the project tests
On macOS the test TestProjectCreate would fail on the final assert that the user is located again in
his original working directory. The test would retrieve the cwd from os.Getwd(). Note, tests that
create files and folders are usually created in the system native temp folder. On macOS that's under
/tmp. The problem is, that /tmp is a symlink that points to /private/tmp. So, when TestProjectCreate
made the final assert that the current working directory is equal to the original working directory
os.Getwd() would return /private/tmp instead of the expected /tmp because the function resolves the
symlink. This test only failed on macOS.

Thanks to Shirley Leu (https://github.com/shirleyleu) for the bug report.