refactor: add support for packages and collections
Proji packages and collections

Packages and collections reinvent the way you import classes into proji.

**What is a package?**
We can already import local configs or the structure of a repository. A package consists of a class config file and the associated templates and scripts, and is hosted in a repo on github or gitlab.
Until now, if you pulled a config from, say, github, you had to take care of dependencies like scripts and templates yourself. If you import a package instead, all dependencies are imported immediately and the class is ready to use. The easiest way to exchange classes.

*Example:*

proji class import --package gh:/nikoksr/proji-test-package/blob/master/configs/c-lang.toml (yes, domain shortcuts are now supported.)

This will import the given config and download all scripts and templates into your main config folder. You don't have to do anything else and can immediately create new C++ projects.

**Collections**
The disadvantage of packages would be that you would have to create a separate repo for each package (class), which is not a problem with today's possibilities but would be very annoying and tedious. And that's where collections come into play. Collections combine multiple packages into a single repo. A collection can contain an almost endless number of packages.
The repo of a collection must contain the three folders 'configs/', 'scripts/' and 'templates/'. You can import a specific package from this collection using `--import package </path/to/package>` or you can import the whole collection using `--import collection </path/to/collection>`. This will import each package and its dependencies from that collection.

Collections also behave relatively smart and try to work as space-saving and efficient as possible. For example, it is relativly common for classes to access the same script, e.g. a script to initialize git. It would be wasteful if each package managed its own dependencies. How would one package know if the script it depends on already exists in the repo? So the collection collects all scripts and templates in dedicated folders, which means that there are virtually no duplicates and packages can share scripts and templates. So a script or a template can be the dependency for a large number of packages.

An example collection hosted on GitHub: https://github.com/nikoksr/proji-test-package

Try it out.

Import for example the python package with:

> proji class import --package gh:/nikoksr/proji-test-package/blob/master/configs/python.toml

Or import the whole collection with:

> proji class import --collection gh:/nikoksr/proji-test-package

Closes #113