refactor: paths in main config may now be relative or absolute
Before this the given path for e.g. the db was always set relative to
the main config folder. So if you set the db path in the main config to
'/db/config.sqlite3' proji would append it to the main configs path and
use the result as path to the sqlite db. If the user would have set the
path to an absolute path, proji would have appended it to the main
configs path anyway, resulting in a 'file not found' error.
Now the given path is checked if it's absolute and gets treated
accordingly. Paths like 'db/proji.sqlite3' will still be appended to the
main config folders path and absolute paths like '/my/custom/path/proji.sqlite3'
will be used as the absolute path to the db.