refactor: move generation of config folder and file to here
deploy now manages the creation of the main config folder and file. It
is called from cmd/init and initializes the main config folder, if it
does not exist.
Additionally it now doesn't download the main config no more but rather
creates a empty config instance, loads it default values and writes it
to memory.