refactor: project create function now reads cwd itself
Until now project.Create received a string which held the
path of the current working directory. That was done to reduce overhead
when creating multiple projects in the same runtime. The problem is that
the function signature looked really confusing. Why do I have to pass my
current working directory to the project.Create function. The parameter
got removed and the function call itself reads in the cwd. So basically
readability > performance in that case.