refactor: optimize project architecture
Because of the previous project architecture we had many problems with
include cycles when dealing with models and remote repositories. The
necessary modifications led to weird looking function signatures and
calls.
Project has now a clean and easier to understand architecture. In case
of the previously called models, I split the type declarations from the
implementations. Every major domain is now described in pkg/domain like
for example package and project. pkg/domain declares the type and its
store and service respectively. The store interface describes the way a
domain talks to a storage provider like a database. The service
interface describes how the implementer would usually interact with the
domain. Some service functions use store functions internally.