refactor(export): replace example with template flag
Instead of exporting a big and descriptive example package config, -t
now exports a minimal package config which serves as a quick entrypoint
for the creation of a new package. The big descriptive config file now
lives in the github repo and will serve as a reference for new users.