refactor(config): add config option import.exclude
The main config now holds a field which lets the user specify a regex
which is used with the 'package import' command. The specified regex is
used to exclude folders and files from imports via dir/repo-structure
and collections. In case that the '-e/--exclude' flag of said command
gets passed, it overwrites the regex specified in the config file.