refactor: add statuswriter package
The statuswriter package is meant to print the satuses of mulitple
actions to the cli concurrently. The statuswriter can hold multiple
sinks. Each sink will write and update a single line in the terminal.
Its main use case currently lies in giving good feedback to the use
while importing multiple packages. Each packages import status will be
written in a dedicated line. The line will be rewritten when the state
of the package changes. E.g. the line will update when the status
swiches from importing the package to storing it.