refactor(store): remove soft-deletes from sql store
Records will no longer be soft-deleted, meaning that a for example
deleted package configuration will no longer be recoverable. This
feature introduced more difficulties than it solved. Soft-deletes might
return in the future but for now I want a stable core for the storage
system and not fight with duplicate soft-deleted records.