refacor(store): replace most of gorm functional queries with raw sql
For small packages the performance that I got out of gorm was very good.
Please consider that at the time writing this gorm 2.0 is just on the
edge of being released and its documentation is not yet at 100%.
In the extreme case of importing something like the structure of the
GitHub repository torvalds/linux gorm showed significant performance
issues. Importing a package of that size with gorm took literally
minutes. And even deleting something of that size and complexity took a
noticeable time.
I replaced almost all storage interactions that are performance critical
with raw sql and importing torvalds/linux now only takes about 5 seconds
on my laptop which is good enough for now. 99% of packages will be a
fraction of that size and complexity.