refactor(package): replace filters with pure excludes
Previously we were using filters to filter the imported paths and
packages. Filter was meant to allow you to say something like 'take only
this package and that package' or 'take everything except this'. Regexp
from the go stdlib guarantees a consistent runtime for all its functions
and thus does not include support for negative lookaheads which would be
needed for 'only-this' filters. Thats why we now will explicitly use the
term excludes to emphasise that the regex is meant to exlude paths and
or packages.