refactor(import): better handling of duplicate packages
When importing a package with a duplicate label (one that already exists
in storage) proji used to just cancel the import. You had to alter the
label of the existing package in order for you to be able to import the
new package.
Now, when proji detects a duplicate label import it will export the
config of the new package into the user cwd. The user can edit the label
in the config to his liking and import the config afterwards.

Closes #138