fix(project): replace project now works as expected
Project replacement was buggy. It was triggered when the root folder of
the project already exists and when a project was already assigned to
this path in the database. Additionally the replacement didn't work even
when triggered correctly.
First of all this shouldn't trigger at all when the folder already
exists, for safety purposes it should just skip the project creation and
print a warning that the folder already exists. This is so that folders
are not overwritten by accident.
And secondly when the folder creation was successful but a project was
already assigned to this path in database and the user wanted to replace
the assigned project with the freshly created one it would fail because
'suddenly' the project folder already exists.