fix(export): fix a bug where package export is missing name
There's currently a bug (https://github.com/pelletier/go-toml/issues/450) in the toml library
proji is using and its affecting our package export. In some cases the
name field of a package is missing and sometimes its duplicate. The
problem is backtrackable to the order command which makes it so packages
exports always look the same and in their intended order. Until the fix
is out we'll just take out the order command. Export works fine; just
isn't ordered anymore as originally wanted.