feat(cmd): add the 'edit' command
Package editing has just become much simpler. Instead of the old and
cumbersome way (as described in issue #137) there is now the 'edit'
command which will export the chosen packages config to a temporary
file, open that file in your systems prefered graphical text editor,
wait for you to edit and save the config file and replace the old
package with the freshly edited one.